changed the name of ArtistProxy to ServiceProxy and defined interface for ServiceProxy and the resolver was attached to it. I deleted the update with no results.
