+++ app/src/main/java/ayds/newyork/songinfo/moredetails/data/repository/externalServiceProxy/ServiceProxy.kt
@@ -1,0 +1,14 @@
+package ayds.newyork.songinfo.moredetails.data.repository.externalServiceProxy
+
+import ayds.newyork.songinfo.moredetails.domain.entities.Card
+
+interface ServiceProxy {
+    fun getCard(artistName: String): Card?
+}
+
+
+
+
+
+
+
